// src/data/models/Player.ts
export interface Player {
  avatar: string
  player_id: number
  url: string
  username: string
  title: string
  followers: number
  country: string
  last_online: number
  joined: number
  status: string
  is_streamer: boolean
  verified: boolean
  league: string
  streaming_platforms: string[]
}